Add redislog.ship_to_redis_only option
When messages are correctly received by Redis there is no reason to
have them managed by others hooks in the logging collector.

The ``redislog.ship_to_redis_only`` option, when is ``true``, makes
the Redislog extension disable the other hooks if the log message is
correctly received by Redis.
When the option is ``false``, which is the default value, the log
messages will be always sent to Redislog and to the other hooks, such as
the CSV-log.