Add redislog.ship_to_redis_only option
When messages are correctly received by Redis, there might be
no reason to write them into the main PostgreSQL log.

If set to ``true``, the main logger facility of PostgreSQL will be
only used for fallback purposes - for example when no Redis service
is available.

By default, ``redislog.ship_to_redis_only`` is set to ``false``,
meaning that events will be logged both in Redis and in the
main log.